+++ .babelrc
@@ -1,5 +1,5 @@
 {
-    "presets": ["es2015", "stage-1"],
+    "presets": ["es2015", "stage-0"],
     "plugins": [
         ["transform-react-jsx", { "pragma": "ssml" }]
     ]
