+++ package.json
@@ -1,29 +1,52 @@
 {
   "name": "alexa-ssml",
   "version": "0.5.0",
-  "description": "JSX for Alexa Skills Kit SSML",
-  "main": "lib/index.js",
-  "jsnext:main": "src/index.js",
-  "scripts": {
-    "build": "rm -rf lib && babel src --out-dir lib",
-    "lint": "eslint src --fix",
-    "*pretest": "npm run lint",
-    "test": "ava --require babel-register **/__tests__/*-test.js",
-    "prepublish": "in-publish && npm run test && npm run build || not-in-publish",
-    "publish:major": "npm version major && npm publish",
-    "publish:minor": "npm version minor && npm publish",
-    "publish:patch": "npm version patch && npm publish",
-    "postpublish": "git push origin master --tags"
-  },
-  "ava": {
-    "babel": "inherit"
-  },
+  "description": "JSX syntax for Speech Synthesis Markup Language (SSML)",
+  "license": "MIT",
+  "contributors": [
+    "Cameron Hunter <hello@cameronhunter.co.uk> (http://cameronhunter.co.uk)",
+    "Nicholas Clawson <nickclaw@gmail.com> (http://nickclaw.com)"
+  ],
   "repository": {
     "type": "git",
     "url": "https://github.com/nickclaw/alexa-ssml"
   },
   "bugs": {
     "url": "http://github.com/nickclaw/alexa-ssml/issues"
+  },
+  "scripts": {
+    "build": "rm -rf build && babel src --out-dir build",
+    "lint": "eslint src",
+    "pretest": "npm run lint",
+    "test": "ava",
+    "prepublish": "in-publish && npm run test && npm run build || not-in-publish",
+    "publish:major": "npm version major && npm publish",
+    "publish:minor": "npm version minor && npm publish",
+    "publish:patch": "npm version patch && npm publish",
+    "postpublish": "git push origin master --tags"
+  },
+  "main": "build/index.js",
+  "jsnext:main": "src/index.js",
+  "dependencies": {
+    "prop-types": "^0.2.0"
+  },
+  "devDependencies": {
+    "ava": "^0.15.2",
+    "babel": "^6.3.26",
+    "babel-cli": "^6.4.5",
+    "babel-eslint": "^4.1.8",
+    "babel-plugin-transform-react-jsx": "^6.4.0",
+    "babel-preset-es2015": "^6.3.13",
+    "babel-preset-stage-1": "^6.3.13",
+    "babel-register": "^6.4.3",
+    "eslint": "^1.10.3",
+    "eslint-config-airbnb": "^4.0.0",
+    "eslint-plugin-react": "^3.16.1",
+    "in-publish": "^2.0.0"
+  },
+  "ava": {
+    "require": "babel-register",
+    "babel": "inherit"
   },
   "keywords": [
     "alexa",
@@ -36,33 +59,5 @@
     "markup",
     "echo",
     "ability"
-  ],
-  "author": {
-    "name": "Nicholas Clawson",
-    "email": "nickclaw@gmail.com",
-    "url": "nickclaw.com"
-  },
-  "license": "MIT",
-  "dependencies": {
-    "lodash": "^4.0.1",
-    "prop-types": "^0.2.0",
-    "tv4": "^1.2.7",
-    "xmlbuilder": "^4.2.1"
-  },
-  "devDependencies": {
-    "ava": "^0.15.2",
-    "babel": "^6.3.26",
-    "babel-cli": "^6.4.5",
-    "babel-eslint": "^4.1.8",
-    "babel-plugin-transform-react-jsx": "^6.4.0",
-    "babel-preset-es2015": "^6.3.13",
-    "babel-preset-stage-1": "^6.3.13",
-    "babel-register": "^6.4.3",
-    "chai": "^3.4.1",
-    "eslint": "^1.10.3",
-    "eslint-config-airbnb": "^4.0.0",
-    "eslint-plugin-react": "^3.16.1",
-    "in-publish": "^2.0.0",
-    "mocha": "^2.4.4"
-  }
+  ]
 }
