+++ src/schema/break.js
@@ -4,7 +4,7 @@ export default {
     tag: 'break',
     propTypes: {
         strength: PropTypes.oneOf(['none', 'x-weak', 'weak', 'medium', 'strong', 'x-strong']),
-        time: PropTypes.match(/^\d+(?:m?s)?$/),
+        time: PropTypes.match(/^\d+m?s$/),
         children: PropTypes.none
     }
 };
