+++ test/schema/say-as-test.js
@@ -1,5 +1,5 @@
 import test from 'ava';
-import ssml from '../../ssml';
+import ssml from '../../src/ssml';
 
 test('children must be present', t => {
     t.throws(() => <say-as interpret-as="characters" />);
